made a maketxt4.py file which stores data into all banks. we take the first input and store into bank 0 index 0, second data bank1 index 0 and so on untill we hit the end of banks. then we write to bank 0 index 1 and so on.
to do this we wrote a file that takes the binary input and writes it as
a string to be concatinated if teh banks run out. i.e. if i have data
that is to be stores to bank 0 index 1, i will write to the entire bank
0 as 0x0104 where 04 is data at bank 0 index 0 and 01 is data at bank 0
index 1.  thus we can populate the data memory of any no of data
elements